Speaker back button now its own component
♻️ refactor: Speaker back button now its own component
